RHOAIENG-9215 Reorganize Existing Segment Tracking Implementation
