+++ frontend/src/concepts/analyticsTracking/segmentKeyService.ts
@@ -1,5 +1,6 @@
 import axios from '~/utilities/axios';
-import { ODHSegmentKey } from '~/types';
+
+import { ODHSegmentKey } from '~/concepts/analyticsTracking/trackingProperties';
 
 export const fetchSegmentKey = (): Promise<ODHSegmentKey> => {
   const url = '/api/segment-key';
