+++ frontend/src/concepts/analyticsTracking/useTrackHistory.ts
