+++ frontend/src/pages/projects/notebook/NotebookRouteLink.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button, ButtonVariant, Flex, FlexItem, Icon, Tooltip } from '@patternfly/react-core';
 import { ExclamationCircleIcon, ExternalLinkAltIcon } from '@patternfly/react-icons';
 import { NotebookKind } from '~/k8sTypes';
-import { fireTrackingEventRaw } from '~/utilities/segmentIOUtils';
+import { fireTrackingEventRaw } from '~/concepts/analyticsTracking/segmentIOUtils';
 import { getDisplayNameFromK8sResource } from '~/concepts/k8s/utils';
 import useRouteForNotebook from './useRouteForNotebook';
 import { hasStopAnnotation } from './utils';
